pyproject.toml
[tool.poetry]
name = "RecTools"
version = "0.12.0"
description = "An easy-to-use Python library for building recommendation systems"
license = "Apache-2.0"
authors = [
    "Emiliy Feldman <feldlime@yandex.ru>",
    "Daria Tikhonovich <daria.m.tikhonovich@gmail.com>",
    "Daniil Potapov <sharth23@gmail.com>",
    "Ildar Safilo <irsafilo@gmail.com>",
    "Alexander Butenko <a.butenko.o@gmail.com>",
    "Julia Karamnova <juliakaramnova@yandex.ru>",
    "Artem Senin <artemseninhse@gmail.com>",
    "Mikhail Khasykov <mkhasykov@gmail.com>",
    "Mike Sokolov <mikesokolovv@gmail.com>",
    "Andrey Semenov <in48semenov@yandex.ru>",
    "Maxim Lukin <maximlukin1@gmail.com>"
]
maintainers = [
    "Emiliy Feldman <feldlime@yandex.ru>",
    "Daria Tikhonovich <daria.m.tikhonovich@gmail.com>",
]
readme = "README.md"
homepage = "https://github.com/MobileTeleSystems/RecTools"
repository = "https://github.com/MobileTeleSystems/RecTools"
documentation = "https://rectools.readthedocs.io"
keywords = [
    "recsys",
    "recommendation systems",
    "machine learning",
    "AI",
    "personalization",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
]
packages = [
    { include = "rectools", from = "." },
]


[tool.poetry.dependencies]
python = ">=3.9, <3.13"
numpy = [
    {version = ">=1.22, <2.0.0", python = "<3.12"},
    {version = ">=1.26, <2.0.0", python = "3.12"}  # numpy <1.26 fails to install on Python 3.12
]
pandas = ">=1.5.0, <3.0.0"
scipy = "^1.10.1, <1.13"  # in 1.13 were introduced significant changes breaking our logic
tqdm = "^4.27.0"
implicit = "^0.7.1"
attrs = ">=19.1.0,<24.0.0"
typeguard = "^4.1.0"
pydantic = "^2.8.2"
pydantic-core = "^2.20.1"
typing-extensions = "^4.12.2"

# The latest released version of lightfm is 1.17 and it's not compatible with PEP-517 installers (like latest poetry versions).
rectools-lightfm = {version = "^1.17.3", optional = true}

nmslib = {version = "^2.0.4", python = "<3.11", optional = true}
# nmslib officialy doens't support Python 3.11 and 3.12. Use https://github.com/metabrainz/nmslib-metabrainz instead
nmslib-metabrainz = {version = "^2.1.3", python = ">=3.11, <3.13", optional = true}

# The latest torch version available for MacOSX + x86_64 is 2.2.2
torch = [
    {version = ">=1.6.0, <2.3.0", markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", optional = true},
    {version = ">=1.6.0, <3.0.0", optional = true}
]
pytorch-lightning = {version = ">=1.6.0, <3.0.0", optional = true}

ipywidgets = {version = ">=7.7,<8.2", optional = true}
plotly = {version="^5.22.0", optional = true}
nbformat = {version = ">=4.2.0", optional = true}
cupy-cuda12x = {version = "^13.3.0", python = "<3.13", optional = true}
catboost = {version = "^1.1.1", optional = true}


[tool.poetry.extras]
lightfm = ["rectools-lightfm"]
nmslib = ["nmslib", "nmslib-metabrainz"]
torch = ["torch", "pytorch-lightning"]
visuals = ["ipywidgets", "plotly", "nbformat"]
cupy = ["cupy-cuda12x"]
catboost = ["catboost"]

all = [
    "rectools-lightfm",
    "nmslib", "nmslib-metabrainz",
    "torch", "pytorch-lightning",
    "ipywidgets", "plotly", "nbformat",
    "cupy-cuda12x",
    "catboost"
]


[tool.poetry.group.dev.dependencies]
black = "24.4.2"
isort = "5.13.2"
pylint = "3.1.0"
mypy = "1.13.0"
flake8 = "7.0.0"
bandit = "1.7.8"
pytest = "8.1.1"
radon = "6.0.1"
coverage = "7.5.0"
autopep8 = "2.1.0"
codespell = "2.2.6"
pytest-subtests = "0.12.1"
flake8-docstrings = "1.7.0"
pep8-naming = "0.13.3"
pytest-cov = "5.0.0"
pytest-mock = "3.14.0"
click = "8.1.7"
gitpython = "3.1.43"


[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
